requirements.txt
langchain
langgraph
langchain-google-genai
langchain-community
supabase
tavily-python
pypdf
python-dotenv